+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/data/source/network/RepoInfo.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check.data.remote.models
+package jp.co.yumemi.android.code_check.data.source.network
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
